fix(scripts): copy mk_ld.sh to NoAxiom/lib/scripts
reason: the kernel compile and build depends on linker scripts,
and mk_ld.sh plays important role for copying .ld files when user tries to build.

Plus, we will include this file for build if we want to publish this in release,
so I prefer to place this in root repo instead of submodule repo.